[mod] Instance list from dalf/searx-instances
