+++ searxstats/source/__init__.py
@@ -1,3 +1,3 @@
-from .searx_docs import SEARX_INSTANCES_URL, get_instance_urls
+from .searx_instances import SEARX_INSTANCES_URL, get_instance_urls
 
 __all__ = ['SEARX_INSTANCES_URL', 'get_instance_urls']
